+++ diventi/blog/translation.py
@@ -1,13 +1,18 @@
 from modeltranslation.translator import register, TranslationOptions
 
-from .models import Category, Article
+from .models import ArticleCategory, Article, BlogCover
 
 
-@register(Category)
-class CategoryTranslationOptions(TranslationOptions):
+@register(ArticleCategory)
+class ArticleCategoryTranslationOptions(TranslationOptions):
     fields = ('title',)
 
 
 @register(Article)
 class ArticleTranslationOptions(TranslationOptions):
     fields = ('title', 'description', 'caption', 'content', 'slug')
+
+
+@register(BlogCover)
+class BlogCoverTranslationOptions(TranslationOptions):
+    fields = ('label',)